:truck: :fire: Remove boilerplate from create-react-app and move App into components directory
